pyproject.toml
[tool.poetry]
name = "soundcloud-scraper"
version = "0.1.0"
description = ""
authors = ["Fabian Koller <github@kthxbye.cyou>"]
readme = "README.md"
packages = [{include = "soundcloud_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.36.0"
nodriver = "^0.36"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
